backend/requirements.txt
Django==5.0.2
mysqlclient==2.2.4
djangorestframework==3.15.0
django-cors-headers==4.3.1
